# Python: Fix caching of FileSystem