--- a/python/tests/io/test_pyarrow.py
+++ b/python/tests/io/test_pyarrow.py
@@ -279,7 +279,7 @@ def test_deleting_s3_file_no_permission() -> None:
     s3fs_mock = MagicMock()
     s3fs_mock.delete_file.side_effect = OSError("AWS Error [code 15]")
 
-    with patch.object(PyArrowFileIO, "_get_fs") as submocked:
+    with patch.object(PyArrowFileIO, "_initialize_fs") as submocked:
         submocked.return_value = s3fs_mock
 
         with pytest.raises(PermissionError) as exc_info:
@@ -294,7 +294,7 @@ def test_deleting_s3_file_not_found() -> None:
     s3fs_mock = MagicMock()
     s3fs_mock.delete_file.side_effect = OSError("Path does not exist")
 
-    with patch.object(PyArrowFileIO, "_get_fs") as submocked:
+    with patch.object(PyArrowFileIO, "_initialize_fs") as submocked:
         submocked.return_value = s3fs_mock
 
         with pytest.raises(FileNotFoundError) as exc_info:
@@ -309,7 +309,7 @@ def test_deleting_hdfs_file_not_found() -> None:
     hdfs_mock = MagicMock()
     hdfs_mock.delete_file.side_effect = OSError("Path does not exist")
 
-    with patch.object(PyArrowFileIO, "_get_fs") as submocked:
+    with patch.object(PyArrowFileIO, "_initialize_fs") as submocked:
         submocked.return_value = hdfs_mock
 
         with pytest.raises(FileNotFoundError) as exc_info: