https://github.com/apache/iceberg/pull/6010 didn't really fix it

Before:
```
python3 /tmp/benchmark.py

import: 0.175605 seconds
catalog: 0.060079 seconds
table: 0.245859 seconds
scan: 5.218688 seconds
Rows: 119890
```

After:
```
python3 /tmp/benchmark.py

import: 0.169811 seconds
catalog: 0.054512 seconds
table: 0.233786 seconds
scan: 2.165105 seconds
Rows: 119890
```